// ui/src/api/generated/model/productListItemResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */

export interface ProductListItemResponse {
  id: number;
  uuid: string;
  product_category_id: number;
  product_category_name: string;
  material_code: string;
  material_description: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
  inspections_count: number;
  inspection_inbound_under_review: number;
  inspection_outbound_under_review: number;
  inspection_inbound_approved: number;
  inspection_outbound_approved: number;
  inspection_inbound_rejected: number;
  inspection_outbound_rejected: number;
}