// ui/src/api/generated/model/productListResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */
import type { ProductListItemResponse } from './productListItemResponse';

export interface ProductListResponse {
  data: ProductListItemResponse[];
  total: number;
  page: number;
  per_page: number;
  total_pages: number;
}